+++ src/components/Pagination.js
@@ -1,44 +1,173 @@
 import React from "react";
+import { css } from "glamor";
+import "./pagination.css";
 
-// class Pagination extends React.Component {
-//     constructor(props){
-//         super(props)
-//     }
+const styles = {
+  list: css({
+    "> li": {
+      display: "inline-block",
+      padding: 10,
+      border: "1px thin gray",
+      margin: 2
+    }
+  }),
+  link: css({
+    cursor: "pointer"
+  })
+};
 
-//   listPages() {
-//     var size = 20;
-//     var items = this.state.data.slice(0, size);
-//     var countTotal = this.state.data.length;
-//     var pages = countTotal / size;
-//     var listPages1;
-//     // return pages;
-//     for (let index = 0; index < pages.length; index++) {
-//       const element = pages[index];
-//       listPages1 += element;
-//     }
-//     return listPages1;
-//   }
+const defaultProps = {
+  initialPage: 1
+};
 
-//   componentDidMount(){
+class Pagination extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { pager: {} };
+  }
 
-//   }
+  componentDidMount() {
+    // set page if items array isn't empty
+    if (this.props.items && this.props.items.length) {
+      this.setPage(this.props.initialPage);
+    }
+  }
 
-//   render() {
-//     return <p>{this.listPages()}</p>;
-//   }
-// }
+  componentDidUpdate(prevProps, prevState) {
+    // reset page if items array has changed
+    if (this.props.items !== prevProps.items) {
+      this.setPage(this.props.initialPage);
+    }
+  }
 
-function Pagination(props) {
-  return (
-    <div>
-      <p>total pages: {props.totalpages}</p>
-      <ul>
-        <li>1</li>
-        <li>2</li>
-        <li>3</li>
+  setPage(page) {
+    var items = this.props.items;
+    var pager = this.state.pager;
+
+    if (page < 1 || page > pager.totalPages) {
+      return;
+    }
+
+    // get new pager object for specified page
+    pager = this.getPager(items.length, page);
+
+    // get new page of items from items array
+    var pageOfItems = items.slice(pager.startIndex, pager.endIndex + 1);
+
+    // update state
+    this.setState({ pager: pager });
+
+    // call change page function in parent component
+    this.props.onChangePage(pageOfItems);
+  }
+
+  getPager(totalItems, currentPage, pageSize) {
+    // default to first page
+    currentPage = currentPage || 1;
+
+    // default page size is 20
+    pageSize = pageSize || 20;
+
+    // calculate total pages
+    var totalPages = Math.ceil(totalItems / pageSize);
+
+    var startPage, endPage;
+    if (totalPages <= 10) {
+      // less than 10 total pages so show all
+      startPage = 1;
+      endPage = totalPages;
+    } else {
+      // more than 10 total pages so calculate start and end pages
+      if (currentPage <= 6) {
+        startPage = 1;
+        endPage = 10;
+      } else if (currentPage + 4 >= totalPages) {
+        startPage = totalPages - 9;
+        endPage = totalPages;
+      } else {
+        startPage = currentPage - 5;
+        endPage = currentPage + 4;
+      }
+    }
+
+    // calculate start and end item indexes
+    var startIndex = (currentPage - 1) * pageSize;
+    var endIndex = Math.min(startIndex + pageSize - 1, totalItems - 1);
+
+    // create an array of pages to ng-repeat in the pager control
+    var pages = [...Array(endPage + 1 - startPage).keys()].map(
+      i => startPage + i
+    );
+
+    // return object with all pager properties required by the view
+    return {
+      totalItems: totalItems,
+      currentPage: currentPage,
+      pageSize: pageSize,
+      totalPages: totalPages,
+      startPage: startPage,
+      endPage: endPage,
+      startIndex: startIndex,
+      endIndex: endIndex,
+      pages: pages
+    };
+  }
+
+  render() {
+    var pager = this.state.pager;
+
+    if (!pager.pages || pager.pages.length <= 1) {
+      // don't display pager if there is only 1 page
+      return null;
+    }
+
+    return (
+      <ul {...styles.list}>
+        <li className={pager.currentPage === 1 ? "disabled" : ""}>
+          <a {...styles.link} onClick={() => this.setPage(1)}>
+            First
+          </a>
+        </li>
+        <li className={pager.currentPage === 1 ? "disabled" : ""}>
+          <a
+            {...styles.link}
+            onClick={() => this.setPage(pager.currentPage - 1)}
+          >
+            Previous
+          </a>
+        </li>
+        {pager.pages.map((page, index) => (
+          <li
+            key={index}
+            className={pager.currentPage === page ? "active" : ""}
+          >
+            <a {...styles.link} onClick={() => this.setPage(page)}>
+              {page}
+            </a>
+          </li>
+        ))}
+        <li
+          className={pager.currentPage === pager.totalPages ? "disabled" : ""}
+        >
+          <a
+            {...styles.link}
+            onClick={() => this.setPage(pager.currentPage + 1)}
+          >
+            Next
+          </a>
+        </li>
+        <li
+          className={pager.currentPage === pager.totalPages ? "disabled" : ""}
+        >
+          <a {...styles.link} onClick={() => this.setPage(pager.totalPages)}>
+            Last
+          </a>
+        </li>
       </ul>
-    </div>
-  );
+    );
+  }
 }
 
+// Pagination.propTypes = propTypes;
+Pagination.defaultProps = defaultProps;
 export default Pagination;
